+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- `Context.get_channel` no longer raises an assertion error if the cache returns `None`.
 
 ## [2.3.0a1] - 2022-01-13
 ### Added